+++ src/utils/exceptionThrowers.ts
@@ -1,0 +1,20 @@
+import ChatSDKErrors from "../core/ChatSDKErrors";
+import ScenarioMarker from "../telemetry/ScenarioMarker";
+import TelemetryEvent from "../telemetry/TelemetryEvent";
+
+export const throwOCSDKInitializationFailure = (e: unknown, scenarioMarker: ScenarioMarker): void => {
+    const exceptionDetails = {
+        response: ChatSDKErrors.OCSDKInitializationFailure,
+        errorObject: `${e}`
+    }
+
+    scenarioMarker.failScenario(TelemetryEvent.InitializeChatSDK, {
+        ExceptionDetails: JSON.stringify(exceptionDetails)
+    });
+
+    throw Error(exceptionDetails.response);
+};
+
+export default {
+    throwOCSDKInitializationFailure
+}
